Bug Fixing, Add languages through Put
Did some tweaks to create so that it will only create new greetings instead using the entered object's id and overriding an existing one. Same logic applied to languages, and now getting the ID of a language wrong will no longer fail the api call. I do this by using the "findByName" repo command to ensure that the language uses the correct ID if it exists. If not, will create a new one.